:star: Reorganize code, improve clarity & modularity
