+++ src/components/dashboardCard.js
@@ -1,5 +1,5 @@
-const dashboardCard = () => {
-  return `
+const dashboardCard = (() =>
+  `
     <div class="card bg-light card-body mb-3">
       <div class="card-header">
         <h3>
@@ -10,7 +10,6 @@ const dashboardCard = () => {
         Not using any yet
       </div>
     </div>
-  `;
-};
+  `)();
 
 export default dashboardCard;
