+++ src/index.js
@@ -1,34 +1,35 @@
-import './css/colorpicker.css';
-// import 'bootstrap-material-design/dist/css/bootstrap-material-design.min.css';
-import './scss/app.scss';
+/**
+ * Entry point for all JavaScript
+ */
 
-// import 'jquery';
-// import 'popper.js';
-import './js/colorpicker';
-// import 'bootstrap-material-design/dist/js/bootstrap-material-design.min.js';
+'use strict';
+
+import './css/colorpicker.css';
+import './scss/app.scss';
 import loginCard from './components/loginCard.js';
 import copyFromCard from './components/copyFromCard.js';
 import managementCard from './components/managementCard.js';
 import loadingModal from './components/loadingModal.js';
 import app from './js/app.js';
 
-const mainContent = () => {
-  return `
-    <div id="content" class="container">
+const mainContent = document.createElement('div');
+
+mainContent.innerHTML = (() =>
+  `
+    <div id="main-content" class="container">
       <div class="row">
         <div class="col-12 col-lg-4">
-          ${loginCard()}
-          ${copyFromCard()}
+          ${loginCard}
+          ${copyFromCard}
         </div>
         <div class="col-12 col-lg-8">
-          ${managementCard()}
+          ${managementCard}
         </div>
       </div>
-    </div><!-- /container -->
-    ${loadingModal()}
-  `;
-};
+    </div><!-- #content -->
+    ${loadingModal}
+  `)();
 
-document.body.innerHTML += mainContent();
+document.body.appendChild(mainContent);
 
 app();
