+++ src/scss/app.scss
@@ -8,7 +8,7 @@
 body {
   background-color: #fcfcfc;
 
-  #content {
+  #main-content {
     margin-top: 1.5rem;
     margin-bottom: 1.5rem;
   }
